# specs/001-content-foundations/contracts/content-api.openapi.yaml
openapi: 3.1.0
info:
  title: MyOwnNotion Canonical Content API
  version: 0.1.0
  description: >-
    Development contract for canonical pages, folders, files, placements,
    relationships (including non-hierarchical internal page links), lifecycle, lineage, and export. This feature does not make
    the API safe for public exposure; authentication is specified separately.
servers:
  - url: http://127.0.0.1:3000
    description: Local development only
...
paths:
...
  /v1/items:
    get:
      tags: [Items]
      operationId: listItems
      parameters:
        - in: query
          name: parentItemId
          schema: { $ref: '#/components/schemas/Uuid' }
        - in: query
          name: lifecycle
          schema: { $ref: '#/components/schemas/Lifecycle' }
      responses:
        '200':
          description: Canonical items and requested placements
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items: { $ref: '#/components/schemas/Item' }
    post:
      tags: [Items]
      operationId: createItem
      parameters:
        - $ref: '#/components/parameters/MutationId'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateItem' }
      responses:
        '201':
          description: Item and initial revision created atomically
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MutationResult' }
        '400': { $ref: '#/components/responses/Problem' }
        '409': { $ref: '#/components/responses/Problem' }
  /v1/items/{itemId}:
    get:
      tags: [Items]
      operationId: getItem
      parameters:
        - $ref: '#/components/parameters/ItemId'
      responses:
        '200':
          description: Item with active placements
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Item' }
        '404': { $ref: '#/components/responses/Problem' }
    patch:
      tags: [Items]
      operationId: updateItem
      parameters:
        - $ref: '#/components/parameters/ItemId'
        - $ref: '#/components/parameters/MutationId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [baseRevisionId]
              properties:
                baseRevisionId: { $ref: '#/components/schemas/Uuid' }
                name: { $ref: '#/components/schemas/DisplayName' }
              additionalProperties: false
      responses:
        '200':
          description: Accepted update
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MutationResult' }
        '400': { $ref: '#/components/responses/Problem' }
        '409': { $ref: '#/components/responses/Problem' }
...
  /v1/items/{itemId}/trash:
    post:
      tags: [Items]
      operationId: trashItem
      parameters:
        - $ref: '#/components/parameters/ItemId'
        - $ref: '#/components/parameters/MutationId'
      responses:
        '200':
          description: Item or branch entered the 30-day trash
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MutationResult' }
        '409': { $ref: '#/components/responses/Problem' }
  /v1/items/{itemId}/restore:
    post:
      tags: [Items]
      operationId: restoreItem
      parameters:
        - $ref: '#/components/parameters/ItemId'
        - $ref: '#/components/parameters/MutationId'
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                fallbackParentItemId:
                  oneOf:
                    - { $ref: '#/components/schemas/Uuid' }
                    - { type: 'null' }
              additionalProperties: false
      responses:
        '200':
          description: Item restored with a valid placement
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MutationResult' }
        '409': { $ref: '#/components/responses/Problem' }
  /v1/items/{itemId}/placements:
    post:
      tags: [Placements]
      operationId: addFilePlacement
      parameters:
        - $ref: '#/components/parameters/ItemId'
        - $ref: '#/components/parameters/MutationId'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreatePlacement' }
      responses:
        '201':
          description: Placement created without copying logical file content
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MutationResult' }
        '400': { $ref: '#/components/responses/Problem' }
        '409': { $ref: '#/components/responses/Problem' }
  /v1/placements/{placementId}/move:
    post:
      tags: [Placements]
      operationId: movePlacement
      parameters:
        - $ref: '#/components/parameters/PlacementId'
        - $ref: '#/components/parameters/MutationId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [parentItemId, positionKey]
              properties:
                parentItemId:
                  oneOf:
                    - { $ref: '#/components/schemas/Uuid' }
                    - { type: 'null' }
                positionKey: { type: string, minLength: 1, maxLength: 255 }
              additionalProperties: false
      responses:
        '200':
          description: Placement or complete branch moved atomically
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MutationResult' }
        '409': { $ref: '#/components/responses/Problem' }
  /v1/placements/{placementId}:
    delete:
      tags: [Placements]
      operationId: removePlacement
      parameters:
        - $ref: '#/components/parameters/PlacementId'
        - $ref: '#/components/parameters/MutationId'
      responses:
        '200':
          description: Placement removed; final file placement also trashes the file
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MutationResult' }
        '409': { $ref: '#/components/responses/Problem' }
...
components:
  parameters:
    ItemId:
      in: path
      name: itemId
      required: true
      schema: { $ref: '#/components/schemas/Uuid' }
    PlacementId:
      in: path
      name: placementId
      required: true
      schema: { $ref: '#/components/schemas/Uuid' }
...
    MutationId:
      in: header
      name: Idempotency-Key
      required: true
      description: Client-generated UUIDv7 mutation identity
      schema: { $ref: '#/components/schemas/Uuid' }
  responses:
    Problem:
      description: Safe validation, conflict, or integrity error
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
  schemas:
    Uuid:
      type: string
      format: uuid
    DisplayName:
      type: string
      minLength: 1
      maxLength: 512
    ItemKind:
      enum: [page, folder, file]
    Lifecycle:
      enum: [active, trashed, purged]
    PlacementKind:
      enum: [hierarchy, attachment]
...
    PageDocument:
      type: object
      required: [format, formatVersion, body]
      properties:
        format: { const: myownnotion.document+json }
        formatVersion: { type: integer, minimum: 1 }
        body: { type: object }
      additionalProperties: false
    Placement:
      type: object
      required: [id, itemId, kind, parentItemId, positionKey]
      properties:
        id: { $ref: '#/components/schemas/Uuid' }
        itemId: { $ref: '#/components/schemas/Uuid' }
        kind: { $ref: '#/components/schemas/PlacementKind' }
        parentItemId:
          oneOf:
            - { $ref: '#/components/schemas/Uuid' }
            - { type: 'null' }
        positionKey: { type: string }
    Item:
      type: object
      required: [id, kind, name, lifecycle, currentRevisionId, placements]
      properties:
        id: { $ref: '#/components/schemas/Uuid' }
        kind: { $ref: '#/components/schemas/ItemKind' }
        name: { $ref: '#/components/schemas/DisplayName' }
        lifecycle: { $ref: '#/components/schemas/Lifecycle' }
        currentRevisionId: { $ref: '#/components/schemas/Uuid' }
        trashedAt: { type: [string, 'null'], format: date-time }
        purgeAfter: { type: [string, 'null'], format: date-time }
        pageDocument:
          oneOf:
            - { $ref: '#/components/schemas/PageDocument' }
            - { type: 'null' }
        placements:
          type: array
          items: { $ref: '#/components/schemas/Placement' }
      allOf:
        - if:
            properties: { kind: { const: page } }
          then:
            required: [pageDocument]
        - if:
            properties: { kind: { enum: [folder, file] } }
          then:
            properties: { pageDocument: { type: 'null' } }
    CreatePlacement:
      type: object
      required: [kind, parentItemId, positionKey]
      properties:
        id:
          allOf:
            - { $ref: '#/components/schemas/Uuid' }
          description: >-
            Optional client-generated UUIDv7 placement identity, so offline
            projections and the canonical store share one placement identity.
        kind: { $ref: '#/components/schemas/PlacementKind' }
        parentItemId:
          oneOf:
            - { $ref: '#/components/schemas/Uuid' }
            - { type: 'null' }
        positionKey: { type: string, minLength: 1, maxLength: 255 }
      additionalProperties: false
    CreateItem:
      type: object
      required: [id, kind, name, placement]
      properties:
        id: { $ref: '#/components/schemas/Uuid' }
        kind: { enum: [page, folder] }
        name: { $ref: '#/components/schemas/DisplayName' }
        placement: { $ref: '#/components/schemas/CreatePlacement' }
        pageDocument: { $ref: '#/components/schemas/PageDocument' }
      additionalProperties: false
      allOf:
        - if:
            properties: { kind: { const: page } }
          then:
            required: [pageDocument]
        - if:
            properties: { kind: { const: folder } }
          then:
            not: { required: [pageDocument] }
...
    MutationResult:
      type: object
      required: [mutationId, revisionIds]
      properties:
        mutationId: { $ref: '#/components/schemas/Uuid' }
        revisionIds:
          type: array
          minItems: 1
          items: { $ref: '#/components/schemas/Uuid' }
        item: { $ref: '#/components/schemas/Item' }
...
    Problem:
      type: object
      required: [type, title, status, code]
      properties:
        type: { type: string, format: uri-reference }
        title: { type: string }
        status: { type: integer, minimum: 400, maximum: 599 }
        code: { type: string }
        detail: { type: string }
        invalidFields:
          type: array
          items:
            type: object
            required: [field, code]
            properties:
              field: { type: string }
              code: { type: string }